// src/app/components/show/show.interface.ts
export interface IShow {
  _id: string;
  showName: any; // tslint:disable-line
  theaterName: any; // tslint:disable-line
  theaterAddress: any; // tslint:disable-line
  theaterPhoneNumber: any; // tslint:disable-line
  playDate: string;
  image: string;
  cover: string;
  createdDate: string;
  entranceTime: string;
  duration: number;
  durationText: string;
  repeat: number;
  minPrice: number;
  maxPrice: number;
  description: any; // tslint:disable-line
  dates: any; // tslint:disable-line
  actors: any; // tslint:disable-line
  directors: any; // tslint:disable-line
  authors: any; // tslint:disable-line
  cast: any; // tslint:disable-line
  country: any; // tslint:disable-line
}